Functional radio storage system
Added the ability to create new radios and define if they should be private or not. Also refactored how the database is handled, now each radio is entered as a single entry to the database. Made some modifications to ydl options to allow webpage url to be pulled. Among other improvements, and bug fixes